Add files via upload
Please check to make sure this runs and produces intelligible graphs.

I think cleaning this script up and generalizing to more than 1 High LET component in a mixture may be all we need to do until  we compute 95% confidence limits for I(d).

Please treat as confidential until our paper on synergy analysis per se (which emphasizes Chromosome aberrations  instead of  HG data) is accepted. If someone stumbles upon it that is OK. But don't advertise this particular script yet.