Add files via upload
This file is designed as an example of how to keep 95%CI calculations for HZE only separate, for the time being, from HZsynergyMain_merge2.R which Edward and I are working on so that the functionality, variable names, commenting etc. is still highly unstable. When these change it impacts Yimin's calculations.