Update drafts.Rmd
Both: Yimin said OK to combining TE and NTE+TE HZE models using an extra FALSE/TRUE. When this is done he will modify his CI calculations accordingly (his figures will not change so he and I are going ahead with that).

Yimin: is this description of your CI calculations correct?

Edward: Good job. Most of my suggested corrections are just wording.  Substantive points are: the MC CI calculation is now done with ~500 parameter sets and the ~500 corresponding so the number of parameter sets is now much smaller than 500*(length(dose vector));  also, we use </b inverse variance weighted \b) nls( ).  Also in 2.1 I thought we should emphasize  that all of the customized programs we devised are available on GitHub.