Update hgData.R
Updated all the data published to date, recalculated standard deviations and the corresponding NWeights (inverse variances) using the binomial method in .93Alp for consistency. Reorganized the data using an auxiliary .csv file which is in the relevant folder here. That .csv file should be double-checked by Claire to try to catch any remaining inconsistencies.