Add files via upload
I think we finally have placeholders for all the the figures of the minor papers and all the main points of the key section, namely Results. I here downloaded today's version of the minor paper and its web supplement. If you need a .pdf instead, please let me know.  Most of the details are incorrect because Edward, Claire, I, and the rest of the pod are still doing lots of cleanup and organizing. But under the messy and chaotic exterior we now have a really solid infrastructure for the paper IMHO. Edward and Claire: please skim the Results section of the paper for orientation while  you are doing cleanup and let me know about major mistakes when you find them. I will stay away from mice for at least a week now while Edward and Claire work their magic. Edward: I hereby sign off on any decisions you will make; the rest of the pod should check with Edward before pushing anything to GitHub.