Add files via upload
Edward: this Friday I would like to discuss the possibility that you can get this particular figure to plot more or less correctly even if many of the other figures are having teething problems. If you can, you should be able to get a corresponding ribbon plot and thereby, as a temporary measure during the next few weeks, check all the essential functionality of the entire code even while big chunks are still giving problems.