Add files via upload
Edward: good that you have probably turned the corner!
The key question is whether we can get any HG graph (from Monte Carlo) that looks like "template for Eq. 3.3.2" in the attached .pdf, where there is (in analogous CA example) a much larger CI in panel B than in panel A. If we can then we are probably finished with the script writing for the minor paper apart from clean-up. The captions and keys need not be cleaned up much because I can do the final details in Adobeillustrator and I can see the essential information in the plots.R itself. If we cannot get such a figuret here are still worries. Can we or can't we get something like "template for Fig. 3.3.2"?

For some reason I seem to be able only to commit this .pdf file and this comment to the main directory, not to the paper subdirectory where it belongs. Should I rename the file MS06_4.13.2018.pdf to indicate that it is later than the .docx version of the paper that is already in the paper subdirectory?