Update plots.R
Edward:  I added one plot that will be used in the paper. I will add the following to the "issues":
an updated .pdf of the paper so one can see where the extra plot goes;
and a .png of the new figure so one can see what kind of horizontal and vertical dimensions I suggest, relative to each other and to the characters (the characters in the key are only for my use; their size never matters).
If you decide to add transparency to clarify the tangle of alpha particle points at low doses great.
In any case all figures will always be fine tuned in Adobe Illustrator eventually so don't worry about fine tuning in the script.
Let me know if commits like this whole commit, arriving sporadically as I fine tune the paper, will work.