Update plots.R
Started to correct  lowLET = TRUE commands (sometimes should be FALSE and vice versa). Found that sometimes when lowLET = TRUE is ok there may be another error: I think the low LET has to come first in various vectors for the command to run properly and assign the right dose to the low LET component. I left the corrections unfinished but eventually they may have to be dealt with. Other problems with plots.R are described in various spots but unfortunately not all in one place. The upshot is that plots.R needs lots of careful work