Update plots.R
Edward: I replaced all of plots.R with a much better version, that is better organized, better commented, and runs nicely. It still needs a lot of work -- for example the lines are too long and some short lines could use explanatory comments which I haven't yet put in. But I plan to leave it as is till we meet next thursday. There is a minor bug in monte_carlo.R ---- uncorr_Bottom and uncorr_Top somehow got switched, but I will leave monte_carlo completely alone for the foreseeable future.