Add files via upload
Edward: I got very valuable feedback from one in-house type reviewer who is an actual biologist-statistician-modeler. 

I couldn't figure out how to incorporate the 9 panel supplement plot into plots.R without having it gum up the works with dev.off(). So I put it into a file parallel to plots.R and put it in the exploratory subfolder. Eventually it will be cloned and put in the repository dedicated to the minor paper.

I put another file needed later but not relevant to us now into the exploratory subfolder instead.